Promote the CGD converter into the package: autografs-topologies
The converter moves from scripts/cgd2pkl.py to autografs.cgd and
installs as the autografs-topologies console command, so regenerating
or extending the topology library no longer requires a repo checkout.
The old script remains as a thin shim.

Along the move:
- TopologyExtractionError joins autografs.exceptions under
  AutografsError.
- The download path drops codecs.escape_decode - a private CPython
  API applied to data that needs no escape decoding (verified: the
  RCSR file parses identically as plain UTF-8) - and gains a timeout
  and raise_for_status.
- tqdm and requests become declared dependencies; both were imported
  while arriving only transitively through pymatgen.
- Missing input files exit with a message instead of an assert.

pyproject also swaps the lint stack in dev extras (ruff replaces
black+isort+flake8; config in [tool.ruff]) and aligns mypy's
python_version with requires-python; the code changes for that land
separately.
